Moved generic code to dlite.mappings
