+++ bindings/python/mappings.py
@@ -1,0 +1,202 @@
+"""Implements mappings between entities.
+
+"""
+from typing import Dict, AnyStr
+
+import dlite
+
+
+class MappingError(Exception):
+    """Base class for mapping errors."""
+
+
+class InsufficientMappingError(MappingError):
+    """There are properties or dimensions that are not mapped."""
+
+
+class AmbiguousMappingError(MappingError):
+    """A property maps to more than one value."""
+
+
+class InconsistentDimensionError(MappingError):
+    """The size of a dimension is assigned to more than one value."""
+
+
+def create_match(triples, match_first=False):
+    """Returns a match functions for `triples`.
+
+    If `match_first` is false, the returned match function will return
+    a generator over all matching triples.  Otherwise the the returned
+    function will only return the first matching triple.
+
+    Args:
+        triples: A list of triples.
+        match_first: Whether the returned match function should return
+          only the first match.
+
+    Example:
+    >>> triples = [
+    ...     (':mamal', 'rdfs:subClassOf', ':animal'),
+    ...     (':cat', 'rdfs:subClassOf', ':mamal'),
+    ...     (':mouse', 'rdfs:subClassOf', ':mamal'),
+    ...     (':cat', ':eats', ':mouse'),
+    ... ]
+    >>> match = create_match(triples)
+    >>> match_first = create_match(triples, only_first=True)
+    >>> list(match(':cat'))
+    [(':cat', 'rdfs:subClassOf', ':mamal'),
+     (':cat', ':eats', ':mouse')]
+    >>> match_first(':cat', None, None)
+    (':cat', 'rdfs:subClassOf', ':mamal')
+    """
+    def match(s=None, p=None, o=None):
+        """Returns generator over all triples that matches (s, p, o)."""
+        return (t for t in triples if
+                (s is None or t[0] == s) and
+                (p is None or t[1] == p) and
+                (o is None or t[2] == o))
+    if match_first:
+        return lambda s=None, p=None, o=None: next(
+            iter(match(s, p, o) or ()), (None, None, None))
+    else:
+        return match
+
+
+def unitconvert_pint(dest_unit, value, unit):
+    """Returns `value` converted to `dest_unit`.
+
+    A unitconvert function based on Pint.  Alternative functions
+    based on ontologies may be implemented.
+
+    Args:
+        dest_unit: Destination unit that `value` should be converted to.
+        value: Source value.
+        unit: The unit of the source value.
+    """
+    import pint
+    ureg = pint.UnitRegistry()
+    u1 = ureg(unit)
+    u2 = ureg(dest_unit)
+    return (value * u1).to(u2).m
+
+
+def assign_dimensions(dims: Dict,
+                      inst,
+                      propname: AnyStr):
+    """Assign dimensions from property assignment.
+
+    Args:
+        dims: A dict with (dimension name, dimension value) pairs that
+          should be assigned.  Only values that are None will be assigned.
+        inst: Source instance.
+        propname: Source property name.
+    """
+    lst = [p.dims for p in inst.meta['properties'] if p.name == propname]
+    if not lst:
+        raise MappingError('Unexpected property name: {src_propname}')
+    src_dims, = lst
+    for dim in src_dims:
+        if dim not in dims:
+            raise InconsistentDimensionError('Unexpected dimension: {dim}')
+        if dims[dim] is None:
+            dims[dim] = inst.dimensions[dim]
+        elif dims[dim] != inst.dimensions[dim]:
+            raise InconsistentDimensionError(
+                f'Trying to assign dimension {dim} of {src_inst.meta.uri} '
+                f'to {src_inst.dimensions[dim]}, but it is already assigned '
+                f'to {dims[dim]}')
+
+
+def make_instance(meta, instances, mappings=(), strict=True,
+                  allow_incomplete=False, unitconvert=unitconvert_pint,
+                  mapsTo=':mapsTo'):
+    """Create an instance of `meta` using data found in `*instances`.
+
+    Args:
+        meta: Metadata for the instance we will create.
+        instances: sequence of instances that the new intance will be
+          populated from.
+        mappings: A sequence of triples defining the mappings.
+        strict: If false, we will allow implicit mapping of properties
+          with the same name.
+        allow_incomplete: Whether to allow not populating all properties
+          of the returned instance.
+        unitconvert: A callable that converts between units.  It has
+          prototype
+
+              unitconvert(dest_unit, value, unit)
+
+          and should return `value` (in units `unit`) converted to
+          `dest_unit`.
+        mapsTo: How the 'mapsTo' predicate is written in `mappings`.
+
+    Returns:
+        New instance of `meta` populated from `*instances`.
+
+    Raises:
+        InsufficientMappingError: There are properties or dimensions in
+          the returned instance that are not mapped.
+        AmbiguousMappingError: A property in the returned instance
+          maps to more than one value.
+        InconsistentDimensionError: The size of a dimension in the
+          returned instance is assigned to more than one value.
+
+    Todo:
+    - Consider that mapsTo is a transitive relation.
+    - Use EMMOntoPy to also account for rdfs:subClassOf relations.
+    - Consider the possibility to assign values via the `mappings`
+      triples.  Do we really want that?  May be useful, but will add
+      a lot of complexity.  Should we have different relations for
+      default values and values that will overwrite what is provided
+      from a matching input instance?
+    """
+    match = create_match(mappings)  # match function
+
+    if isinstance(instances, dlite.Instance):
+        instances = [instances]
+
+    dims = {d.name: None for d in meta['dimensions']}
+    props = {}
+
+    for prop in meta['properties']:
+        prop_uri = f'{meta.uri}#{prop.name}'
+        for _, _, o in match(prop_uri, mapsTo, None):
+            for inst in instances:
+                for prop2 in inst.meta['properties']:
+                    prop2_uri = f'{inst.meta.uri}#{prop2.name}'
+                    for _, _, o2 in match(prop2_uri, mapsTo, o):
+                        value = inst[prop2.name]
+                        if prop.name not in props:
+                            assign_dimensions(dims, inst, prop2.name)
+                            props[prop.name] = value
+                        elif props[prop.name] != value:
+                            raise AmbiguousMappingError(
+                                f'"{prop.name}" maps to both '
+                                f'"{props[prop.name]}" and "{value}"')
+
+        if prop.name not in props and not strict:
+            for inst in instances:
+                if prop.name in inst.properties:
+                    value = inst[prop.name]
+                    if prop.name not in props:
+                        assign_dimensions(dims, inst, prop.name)
+                        props[prop.name] = value
+                    elif props[prop.name] != value:
+                        raise AmbiguousMappingError(
+                            f'"{prop.name}" assigned to both '
+                            f'"{props[prop.name]}" and "{value}"')
+
+        if not allow_incomplete and prop.name not in props:
+            raise InsufficientMappingError(
+                f'no mapping for assigning property "{prop.name}" '
+                f'in {meta.uri}')
+
+    if None in dims:
+        dimname = [k for k, v in dims.items() if v is None][0]
+        raise InsufficientMappingError(
+            f'dimension "{dimname}" is not assigned')
+
+    inst = meta(list(dims.values()))
+    for k, v in props.items():
+        inst[k] = v
+    return inst
